examples/template/rss-twitter-agent/requirements.txt
feedparser>=6.0.0
pyyaml>=6.0
